BF: force int output for math functions
math.ceil, math.floor, round generate integers in Python 3 but not
Python 2.

+++ nipy/algorithms/clustering/imm.py
@@ -236,7 +236,7 @@ class IMM(BGMM):
         if np.isscalar(kfold):
             aux = np.argsort(np.random.rand(n_samples))
             idx = - np.ones(n_samples).astype(np.int)
-            j = math.ceil(n_samples / kfold)
+            j = int(math.ceil(n_samples / kfold))
             kmax = kfold
             for k in range(kmax):
                 idx[aux[k * j:min(n_samples, j * (k + 1))]] = k
